feat(core): add client callbacks
This will add the ability to register callbacks on the client side, then trigger them from the server. It is a nifty thing I made a while ago, but never really put it anywhere. This is especially helpful when you need instant information from a particular client on the server side. 

A good use example of this is say that you would like information about a vehicle the player is in on the server side, you could simply register the client callback for it, then trigger it from the server

```lua
exports.drp_core:RegisterClientCallback('drp:getveh', function(cb)
    local veh = GetVehiclePedIsIn(PlayerPedId())
    cb(GetVehicleSteeringAngle(veh))
end)
```

Then on the server

```lua
RegisterCommand('geit', function(source, args, raw)
    local target = table.unpack(args)
    target = tonumber(target)
    exports.drp_core:TriggerClientCallback(1, 'drp:getveh', function(angle)
    print('the angle was', angle)
end)
end)
```

Types for the exports have been created in `types.lua`. This file should not be loaded in the `fxmanifest` as it is just for editor intellisense.

This feature isn't guaranteed to be extremely performant and may have flaws. Please review and test, I think this is something many could find useful for various things in the future.